remove code smells and bugs V2

+++ core/src/se2/groupb/monopoly/screens/MonopolyScreen.java
@@ -235,8 +235,8 @@ public class MonopolyScreen extends GameScreenAdapter {
         if (player4 != null) {
             moneyfont.draw(batch, player4.getName() + ": " + player4.getBankBalance(), 0, Gdx.graphics.getHeight() - 250);
         }
-        moneyfont.draw(batch, screenOutput, (float) (Gdx.graphics.getWidth() / 3.75), (float) (yPosInitialButtons + 250));
-        moneyfont.draw(batch, "Pot: " + moneyPot.getAmount(), 0, (float) (Gdx.graphics.getHeight() - 400));
+        moneyfont.draw(batch, screenOutput, (float) (Gdx.graphics.getWidth() / 3.75), yPosInitialButtons + 250f);
+        moneyfont.draw(batch, "Pot: " + moneyPot.getAmount(), 0, Gdx.graphics.getHeight() - 400f);
 
         batch.end();
     }
